Move 'P384KeyPair' implementation to its own module.
